move the ciphertext generator to a separate module
